Fixed the label positions str/int bug after it was discovered to exist in the load study script as well and the submission view.
Details:

- Made init_loader in load_table take an optional AggregatedErrors object so that it could incorporate errors/warnings that occurred before instantiation, specifically errors associated with the pandas operations that read the file.
- Read each sheet individually in load_study, to incorporate the dtype argument.
- Redid the changes from the previous PR, because I re-discovered the method "get_study_dtypes_dict()" that I had previously created, which generates the df_dict.  Its docstring also points out that read_from_file does in fact handle the dict of dicts of dtypes, which I had wondered about before, but handle dug into.  That method buffers exceptions similar to how I had done it in the last PR, but it uses the saved loader objects instead of the class method _get_column_types for each loader.  That wan't possible in load_study, because I didn't have access to that method.  I should probably consolidate these methods into one spot in a future refactor.

Files checked in: DataRepo/management/commands/load_study.py DataRepo/management/commands/load_table.py DataRepo/views/upload/submission.py
