+++ DataRepo/views/upload/submission.py
@@ -654,7 +654,7 @@ class BuildSubmissionView(FormView):
             Raises:
                 None
         Returns:
-            df [Union[DataFrame, Dict[str, DataFrame]]
+            df (Union[DataFrame, Dict[str, DataFrame]])
         """
         try:
             df = read_from_file(file, sheet=sheet, dtype=dtype)
@@ -3357,22 +3357,9 @@ class BuildSubmissionView(FormView):
         try:
             # Generate the df_dict, accounting for the data types by supplying the dtype argument.  We will use
             # optional_mode to avoid pandas errors about types that do not allow empty values.
-            aes = AggregatedErrors()
-            df_dict = {}
-            sheets = get_sheet_names(self.study_file)
-            for ldr in StudyLoader.Loaders._asdict().values():
-                if (
-                    ldr is not None
-                    and issubclass(ldr, TableLoader)
-                    and ldr.DataSheetName in sheets
-                ):
-                    ldr_dtypes, ldr_aes = ldr._get_column_types(optional_mode=True)
-                    aes.merge_aggregated_errors_object(ldr_aes)
-
-                    # Get the StudyLoader for the version of the input file
-                    df_dict[ldr.DataSheetName] = self.read_from_file(
-                        self.study_file, sheet=ldr.DataSheetName, dtype=ldr_dtypes
-                    )
+            df_dict = self.read_from_file(
+                self.study_file, sheet=None, dtype=self.get_study_dtypes_dict()
+            )
 
             loader_class = StudyLoader.get_derived_class(df_dict)
             sl: StudyLoader = loader_class(
@@ -3383,8 +3370,6 @@ class BuildSubmissionView(FormView):
                 annot_files_dict=self.annot_files_dict,
             )
             sl.load_data()
-
-            load_status_data.set_load_exception(aes, StudyLoader.ConversionHeading)
         except (
             InvalidStudyDocVersion,
             UnknownStudyDocVersion,
